+++ resources/lang/en/error_catalogue/messages.php
@@ -8,4 +8,5 @@
     'unknown_error' => 'An unknown error occured.',
     'record_not_found' => 'The record was not found.',
     'validation_error' => 'Invalid data provided.',
+    'too_many_requests_error' => 'Too many requests.',
 ];